feat(core): add Engine::google_sheets() conformance factory
